+++ src/dashboard/apigateway/apigateway/apis/web/constants.py
@@ -17,6 +17,8 @@
 #
 from blue_krill.data_types.enum import EnumField, StructuredEnum
 
+from apigateway.core.constants import BackendTypeEnum
+
 
 class UserAuthTypeEnum(StructuredEnum):
     IEOD = EnumField("ieod")
@@ -31,3 +33,29 @@ class ExportTypeEnum(StructuredEnum):
     FILTERED = EnumField("filtered")
     # 已选资源
     SELECTED = EnumField("selected")
+
+
+class BackendConfigTypeEnum(StructuredEnum):
+    NODE = EnumField("node", label="节点")
+    # NOTE: 当前不支持
+    SERVICE_DISCOVERY = EnumField("service_discovery", label="服务发现")
+
+
+class LoadBalanceTypeEnum(StructuredEnum):
+    RR = EnumField("roundrobin", "RR")
+    WRR = EnumField("weighted-roundrobin", "Weighted-RR")
+
+
+class BackendConfigSchemeEnum(StructuredEnum):
+    # for http type backend
+    HTTP = EnumField("http", label="HTTP")
+    HTTPS = EnumField("https", label="HTTPS")
+    # for grpc type backend
+    GRPC = EnumField("grpc", label="GRPC")
+    GRPCS = EnumField("grpcs", label="GRPCS")
+
+
+BACKEND_CONFIG_SCHEME_MAP = {
+    BackendTypeEnum.HTTP.value: [BackendConfigSchemeEnum.HTTP.value, BackendConfigSchemeEnum.HTTPS.value],
+    BackendTypeEnum.GRPC.value: [BackendConfigSchemeEnum.GRPC.value, BackendConfigSchemeEnum.GRPCS.value],
+}
