+++ src/dashboard/apigateway/apigateway/apis/web/stage/serializers.py
@@ -23,8 +23,8 @@ from rest_framework import serializers
 from rest_framework.validators import UniqueTogetherValidator
 from tencent_apigateway_common.i18n.field import SerializerTranslatedField
 
-from apigateway.apis.web.backend.constants import BACKEND_CONFIG_SCHEME_MAP
-from apigateway.apis.web.backend.serializers import BaseBackendConfigSLZ
+from apigateway.apis.web.constants import BACKEND_CONFIG_SCHEME_MAP
+from apigateway.apis.web.serializers import BaseBackendConfigSLZ
 from apigateway.biz.validators import MaxCountPerGatewayValidator
 from apigateway.common.fields import CurrentGatewayDefault
 from apigateway.core.constants import STAGE_NAME_PATTERN, ReleaseStatusEnum, StageStatusEnum
